Update test modules names.
