+++ compiler/qsc_passes/src/capabilitiesck/tests_adaptive.rs
@@ -3,7 +3,7 @@
 
 #![allow(clippy::needless_raw_string_hashes)]
 
-use super::common::{
+use super::tests_common::{
     check, CALL_DYNAMIC_FUNCTION, CALL_DYNAMIC_OPERATION,
     CALL_TO_CICLYC_FUNCTION_WITH_CLASSICAL_ARGUMENT, CALL_TO_CICLYC_FUNCTION_WITH_DYNAMIC_ARGUMENT,
     CALL_TO_CICLYC_OPERATION_WITH_CLASSICAL_ARGUMENT,
@@ -18,7 +18,7 @@ use expect_test::{expect, Expect};
 use qsc_frontend::compile::RuntimeCapabilityFlags;
 
 fn check_profile(source: &str, expect: &Expect) {
-    check(source, expect, RuntimeCapabilityFlags::empty());
+    check(source, expect, RuntimeCapabilityFlags::ForwardBranching);
 }
 
 #[test]
@@ -32,34 +32,21 @@ fn minimal_program_yields_no_errors() {
 }
 
 #[test]
-fn use_of_dynamic_boolean_yields_error() {
+fn use_of_dynamic_boolean_yields_no_errors() {
     check_profile(
         USE_DYNAMIC_BOOLEAN,
         &expect![[r#"
-            [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 104,
-                        hi: 116,
-                    },
-                ),
-            ]
+            []
         "#]],
     );
 }
 
 #[test]
-fn use_of_dynamic_int_yields_errors() {
+fn use_of_dynamic_int_yields_error() {
     check_profile(
         USE_DYNAMIC_INT,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 246,
-                        hi: 271,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 246,
@@ -72,18 +59,15 @@ fn use_of_dynamic_int_yields_errors() {
 }
 
 #[test]
-fn use_of_dynamic_pauli_yields_errors() {
-    // In the case of if expressions, if either the condition or the blocks yield errors, the errors yielded by the
-    // whole if expression are not surfaced to avoid too error churn.
-    // For this reason, the "use of dynamic pauli" error is not yielded in this case.
+fn use_of_dynamic_pauli_yields_error() {
     check_profile(
         USE_DYNAMIC_PAULI,
         &expect![[r#"
             [
-                UseOfDynamicBool(
+                UseOfDynamicPauli(
                     Span {
                         lo: 104,
-                        hi: 116,
+                        hi: 134,
                     },
                 ),
             ]
@@ -97,12 +81,6 @@ fn use_of_dynamic_range_yields_errors() {
         USE_DYNAMIC_RANGE,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 108,
-                        hi: 137,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 108,
@@ -126,12 +104,6 @@ fn use_of_dynamic_double_yields_errors() {
         USE_DYNAMIC_DOUBLE,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 246,
-                        hi: 284,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 246,
@@ -155,12 +127,6 @@ fn use_of_dynamic_qubit_yields_errors() {
         USE_DYNAMIC_QUBIT,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 105,
-                        hi: 123,
-                    },
-                ),
                 UseOfDynamicQubit(
                     Span {
                         lo: 146,
@@ -178,12 +144,6 @@ fn use_of_dynamic_big_int_yields_errors() {
         USE_DYNAMIC_BIG_INT,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 247,
-                        hi: 285,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 247,
@@ -207,12 +167,6 @@ fn use_of_dynamic_string_yields_errors() {
         USE_DYNAMIC_STRING,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 130,
-                        hi: 144,
-                    },
-                ),
                 UseOfDynamicString(
                     Span {
                         lo: 130,
@@ -225,17 +179,11 @@ fn use_of_dynamic_string_yields_errors() {
 }
 
 #[test]
-fn use_of_dynamically_sized_array_yields_errors() {
+fn use_of_dynamically_sized_array_yields_error() {
     check_profile(
         USE_DYNAMICALLY_SIZED_ARRAY,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 104,
-                        hi: 136,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 104,
@@ -259,12 +207,6 @@ fn use_of_dynamic_udt_yields_errors() {
         USE_DYNAMIC_UDT,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 283,
-                        hi: 335,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 283,
@@ -290,17 +232,14 @@ fn use_of_dynamic_udt_yields_errors() {
 
 #[test]
 fn use_of_dynamic_function_yields_errors() {
-    // In the case of if expressions, if either the condition or the blocks yield errors, the errors yielded by the
-    // whole if expression are not surfaced to avoid too much error churn.
-    // For this reason, the "use of dynamic function" error is not yielded in this case.
     check_profile(
         USE_DYNAMIC_FUNCTION,
         &expect![[r#"
             [
-                UseOfDynamicBool(
+                UseOfDynamicArrowFunction(
                     Span {
                         lo: 142,
-                        hi: 154,
+                        hi: 166,
                     },
                 ),
             ]
@@ -310,17 +249,14 @@ fn use_of_dynamic_function_yields_errors() {
 
 #[test]
 fn use_of_dynamic_operation_yields_errors() {
-    // In the case of if expressions, if either the condition or the blocks yield errors, the errors yielded by the
-    // whole if expression are not surfaced to avoid too much error churn.
-    // For this reason, the "use of dynamic operation" error is not yielded in this case.
     check_profile(
         USE_DYNAMIC_OPERATION,
         &expect![[r#"
             [
-                UseOfDynamicBool(
+                UseOfDynamicArrowOperation(
                     Span {
                         lo: 142,
-                        hi: 154,
+                        hi: 162,
                     },
                 ),
             ]
@@ -344,12 +280,6 @@ fn call_cyclic_function_with_dynamic_argument_yields_errors() {
         CALL_TO_CICLYC_FUNCTION_WITH_DYNAMIC_ARGUMENT,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 211,
-                        hi: 243,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 211,
@@ -408,12 +338,6 @@ fn call_cyclic_operation_with_dynamic_argument_yields_errors() {
                         hi: 23,
                     },
                 ),
-                UseOfDynamicBool(
-                    Span {
-                        lo: 212,
-                        hi: 244,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 212,
@@ -437,16 +361,10 @@ fn call_to_dynamic_function_yields_errors() {
         CALL_DYNAMIC_FUNCTION,
         &expect![[r#"
             [
-                UseOfDynamicBool(
+                UseOfDynamicArrowFunction(
                     Span {
                         lo: 142,
-                        hi: 154,
-                    },
-                ),
-                UseOfDynamicBool(
-                    Span {
-                        lo: 180,
-                        hi: 188,
+                        hi: 166,
                     },
                 ),
                 UseOfDynamicDouble(
@@ -478,16 +396,10 @@ fn call_to_dynamic_operation_yields_errors() {
         CALL_DYNAMIC_OPERATION,
         &expect![[r#"
             [
-                UseOfDynamicBool(
+                UseOfDynamicArrowOperation(
                     Span {
                         lo: 142,
-                        hi: 154,
-                    },
-                ),
-                UseOfDynamicBool(
-                    Span {
-                        lo: 176,
-                        hi: 181,
+                        hi: 162,
                     },
                 ),
                 UseOfDynamicArrowOperation(
@@ -531,24 +443,11 @@ fn call_to_unresolved_yields_errors() {
 }
 
 #[test]
-fn measurement_within_dynamic_scope_yields_errors() {
+fn measurement_within_dynamic_scope_yields_no_errors() {
     check_profile(
         MEASUREMENT_WITHIN_DYNAMIC_SCOPE,
         &expect![[r#"
-            [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 99,
-                        hi: 110,
-                    },
-                ),
-                MeasurementWithinDynamicScope(
-                    Span {
-                        lo: 137,
-                        hi: 141,
-                    },
-                ),
-            ]
+            []
         "#]],
     );
 }
@@ -559,22 +458,10 @@ fn use_of_dynamic_index_yields_errors() {
         USE_DYNAMIC_INDEX,
         &expect![[r#"
             [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 246,
-                        hi: 271,
-                    },
-                ),
                 UseOfDynamicInt(
                     Span {
                         lo: 246,
                         hi: 271,
-                    },
-                ),
-                UseOfDynamicBool(
-                    Span {
-                        lo: 319,
-                        hi: 323,
                     },
                 ),
                 UseOfDynamicInt(
@@ -595,24 +482,11 @@ fn use_of_dynamic_index_yields_errors() {
 }
 
 #[test]
-fn return_within_dynamic_scope_yields_errors() {
+fn return_within_dynamic_scope_yields_no_errors() {
     check_profile(
         RETURN_WITHIN_DYNAMIC_SCOPE,
         &expect![[r#"
-            [
-                UseOfDynamicBool(
-                    Span {
-                        lo: 98,
-                        hi: 109,
-                    },
-                ),
-                ReturnWithinDynamicScope(
-                    Span {
-                        lo: 128,
-                        hi: 136,
-                    },
-                ),
-            ]
+            []
         "#]],
     );
 }
@@ -623,16 +497,10 @@ fn loop_with_dynamic_condition_yields_errors() {
         LOOP_WITH_DYNAMIC_CONDITION,
         &expect![[r#"
             [
-                UseOfDynamicBool(
+                UseOfDynamicInt(
                     Span {
                         lo: 106,
-                        hi: 118,
-                    },
-                ),
-                UseOfDynamicBool(
-                    Span {
-                        lo: 141,
-                        hi: 159,
+                        hi: 127,
                     },
                 ),
                 UseOfDynamicInt(
@@ -651,12 +519,6 @@ fn loop_with_dynamic_condition_yields_errors() {
                     Span {
                         lo: 141,
                         hi: 159,
-                    },
-                ),
-                UseOfDynamicBool(
-                    Span {
-                        lo: 150,
-                        hi: 156,
                     },
                 ),
                 UseOfDynamicInt(
