+++ compiler/qsc_passes/src/capabilitiesck/tests_adaptive_plus_integers.rs
@@ -3,7 +3,9 @@
 
 #![allow(clippy::needless_raw_string_hashes)]
 
-use super::common::{
+use crate::capabilitiesck::tests_common::USE_DYNAMIC_RANGE;
+
+use super::tests_common::{
     check, CALL_DYNAMIC_FUNCTION, CALL_DYNAMIC_OPERATION,
     CALL_TO_CICLYC_FUNCTION_WITH_CLASSICAL_ARGUMENT, CALL_TO_CICLYC_FUNCTION_WITH_DYNAMIC_ARGUMENT,
     CALL_TO_CICLYC_OPERATION_WITH_CLASSICAL_ARGUMENT,
@@ -12,13 +14,17 @@ use super::common::{
     RETURN_WITHIN_DYNAMIC_SCOPE, USE_CLOSURE_FUNCTION, USE_DYNAMICALLY_SIZED_ARRAY,
     USE_DYNAMIC_BIG_INT, USE_DYNAMIC_BOOLEAN, USE_DYNAMIC_DOUBLE, USE_DYNAMIC_FUNCTION,
     USE_DYNAMIC_INDEX, USE_DYNAMIC_INT, USE_DYNAMIC_OPERATION, USE_DYNAMIC_PAULI,
-    USE_DYNAMIC_QUBIT, USE_DYNAMIC_RANGE, USE_DYNAMIC_STRING, USE_DYNAMIC_UDT,
+    USE_DYNAMIC_QUBIT, USE_DYNAMIC_STRING, USE_DYNAMIC_UDT,
 };
 use expect_test::{expect, Expect};
 use qsc_frontend::compile::RuntimeCapabilityFlags;
 
 fn check_profile(source: &str, expect: &Expect) {
-    check(source, expect, RuntimeCapabilityFlags::ForwardBranching);
+    check(
+        source,
+        expect,
+        RuntimeCapabilityFlags::ForwardBranching | RuntimeCapabilityFlags::IntegerComputations,
+    );
 }
 
 #[test]
@@ -42,74 +48,41 @@ fn use_of_dynamic_boolean_yields_no_errors() {
 }
 
 #[test]
-fn use_of_dynamic_int_yields_error() {
+fn use_of_dynamic_int_yields_no_errors() {
     check_profile(
         USE_DYNAMIC_INT,
         &expect![[r#"
-            [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 246,
-                        hi: 271,
-                    },
-                ),
-            ]
+            []
         "#]],
     );
 }
 
 #[test]
-fn use_of_dynamic_pauli_yields_error() {
+fn use_of_dynamic_pauli_yields_no_errors() {
     check_profile(
         USE_DYNAMIC_PAULI,
         &expect![[r#"
-            [
-                UseOfDynamicPauli(
-                    Span {
-                        lo: 104,
-                        hi: 134,
-                    },
-                ),
-            ]
+            []
         "#]],
     );
 }
 
 #[test]
-fn use_of_dynamic_range_yields_errors() {
+fn use_of_dynamic_range_yields_no_errors() {
     check_profile(
         USE_DYNAMIC_RANGE,
         &expect![[r#"
-            [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 108,
-                        hi: 137,
-                    },
-                ),
-                UseOfDynamicRange(
-                    Span {
-                        lo: 108,
-                        hi: 137,
-                    },
-                ),
-            ]
+            []
         "#]],
     );
 }
 
 #[test]
-fn use_of_dynamic_double_yields_errors() {
+fn use_of_dynamic_double_yields_error() {
     check_profile(
         USE_DYNAMIC_DOUBLE,
         &expect![[r#"
             [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 246,
-                        hi: 284,
-                    },
-                ),
                 UseOfDynamicDouble(
                     Span {
                         lo: 246,
@@ -144,12 +117,6 @@ fn use_of_dynamic_big_int_yields_errors() {
         USE_DYNAMIC_BIG_INT,
         &expect![[r#"
             [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 247,
-                        hi: 285,
-                    },
-                ),
                 UseOfDynamicBigInt(
                     Span {
                         lo: 247,
@@ -184,12 +151,6 @@ fn use_of_dynamically_sized_array_yields_error() {
         USE_DYNAMICALLY_SIZED_ARRAY,
         &expect![[r#"
             [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 104,
-                        hi: 136,
-                    },
-                ),
                 UseOfDynamicallySizedArray(
                     Span {
                         lo: 104,
@@ -207,12 +168,6 @@ fn use_of_dynamic_udt_yields_errors() {
         USE_DYNAMIC_UDT,
         &expect![[r#"
             [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 283,
-                        hi: 335,
-                    },
-                ),
                 UseOfDynamicDouble(
                     Span {
                         lo: 283,
@@ -275,17 +230,11 @@ fn call_cyclic_function_with_classical_argument_yields_no_errors() {
 }
 
 #[test]
-fn call_cyclic_function_with_dynamic_argument_yields_errors() {
+fn call_cyclic_function_with_dynamic_argument_yields_error() {
     check_profile(
         CALL_TO_CICLYC_FUNCTION_WITH_DYNAMIC_ARGUMENT,
         &expect![[r#"
             [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 211,
-                        hi: 243,
-                    },
-                ),
                 CallToCyclicFunctionWithDynamicArg(
                     Span {
                         lo: 211,
@@ -309,12 +258,6 @@ fn call_cyclic_operation_with_classical_argument_yields_errors() {
                         hi: 23,
                     },
                 ),
-                UseOfDynamicInt(
-                    Span {
-                        lo: 187,
-                        hi: 199,
-                    },
-                ),
                 CallToCyclicOperation(
                     Span {
                         lo: 187,
@@ -336,12 +279,6 @@ fn call_cyclic_operation_with_dynamic_argument_yields_errors() {
                     Span {
                         lo: 15,
                         hi: 23,
-                    },
-                ),
-                UseOfDynamicInt(
-                    Span {
-                        lo: 212,
-                        hi: 244,
                     },
                 ),
                 CallToCyclicOperation(
@@ -458,18 +395,6 @@ fn use_of_dynamic_index_yields_errors() {
         USE_DYNAMIC_INDEX,
         &expect![[r#"
             [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 246,
-                        hi: 271,
-                    },
-                ),
-                UseOfDynamicInt(
-                    Span {
-                        lo: 319,
-                        hi: 323,
-                    },
-                ),
                 UseOfDynamicIndex(
                     Span {
                         lo: 319,
@@ -497,40 +422,10 @@ fn loop_with_dynamic_condition_yields_errors() {
         LOOP_WITH_DYNAMIC_CONDITION,
         &expect![[r#"
             [
-                UseOfDynamicInt(
-                    Span {
-                        lo: 106,
-                        hi: 127,
-                    },
-                ),
-                UseOfDynamicInt(
-                    Span {
-                        lo: 141,
-                        hi: 159,
-                    },
-                ),
-                UseOfDynamicRange(
-                    Span {
-                        lo: 141,
-                        hi: 159,
-                    },
-                ),
                 LoopWithDynamicCondition(
                     Span {
                         lo: 141,
                         hi: 159,
-                    },
-                ),
-                UseOfDynamicInt(
-                    Span {
-                        lo: 150,
-                        hi: 156,
-                    },
-                ),
-                UseOfDynamicRange(
-                    Span {
-                        lo: 150,
-                        hi: 156,
                     },
                 ),
             ]
