+++ compiler/qsc_passes/src/capabilitiesck/tests_base.rs
@@ -3,9 +3,7 @@
 
 #![allow(clippy::needless_raw_string_hashes)]
 
-use crate::capabilitiesck::common::USE_DYNAMIC_RANGE;
-
-use super::common::{
+use super::tests_common::{
     check, CALL_DYNAMIC_FUNCTION, CALL_DYNAMIC_OPERATION,
     CALL_TO_CICLYC_FUNCTION_WITH_CLASSICAL_ARGUMENT, CALL_TO_CICLYC_FUNCTION_WITH_DYNAMIC_ARGUMENT,
     CALL_TO_CICLYC_OPERATION_WITH_CLASSICAL_ARGUMENT,
@@ -14,17 +12,13 @@ use super::common::{
     RETURN_WITHIN_DYNAMIC_SCOPE, USE_CLOSURE_FUNCTION, USE_DYNAMICALLY_SIZED_ARRAY,
     USE_DYNAMIC_BIG_INT, USE_DYNAMIC_BOOLEAN, USE_DYNAMIC_DOUBLE, USE_DYNAMIC_FUNCTION,
     USE_DYNAMIC_INDEX, USE_DYNAMIC_INT, USE_DYNAMIC_OPERATION, USE_DYNAMIC_PAULI,
-    USE_DYNAMIC_QUBIT, USE_DYNAMIC_STRING, USE_DYNAMIC_UDT,
+    USE_DYNAMIC_QUBIT, USE_DYNAMIC_RANGE, USE_DYNAMIC_STRING, USE_DYNAMIC_UDT,
 };
 use expect_test::{expect, Expect};
 use qsc_frontend::compile::RuntimeCapabilityFlags;
 
 fn check_profile(source: &str, expect: &Expect) {
-    check(
-        source,
-        expect,
-        RuntimeCapabilityFlags::ForwardBranching | RuntimeCapabilityFlags::IntegerComputations,
-    );
+    check(source, expect, RuntimeCapabilityFlags::empty());
 }
 
 #[test]
@@ -38,51 +32,112 @@ fn minimal_program_yields_no_errors() {
 }
 
 #[test]
-fn use_of_dynamic_boolean_yields_no_errors() {
+fn use_of_dynamic_boolean_yields_error() {
     check_profile(
         USE_DYNAMIC_BOOLEAN,
         &expect![[r#"
-            []
+            [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 104,
+                        hi: 116,
+                    },
+                ),
+            ]
         "#]],
     );
 }
 
 #[test]
-fn use_of_dynamic_int_yields_no_errors() {
+fn use_of_dynamic_int_yields_errors() {
     check_profile(
         USE_DYNAMIC_INT,
         &expect![[r#"
-            []
+            [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 246,
+                        hi: 271,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 246,
+                        hi: 271,
+                    },
+                ),
+            ]
         "#]],
     );
 }
 
 #[test]
-fn use_of_dynamic_pauli_yields_no_errors() {
+fn use_of_dynamic_pauli_yields_errors() {
+    // In the case of if expressions, if either the condition or the blocks yield errors, the errors yielded by the
+    // whole if expression are not surfaced to avoid too error churn.
+    // For this reason, the "use of dynamic pauli" error is not yielded in this case.
     check_profile(
         USE_DYNAMIC_PAULI,
         &expect![[r#"
-            []
+            [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 104,
+                        hi: 116,
+                    },
+                ),
+            ]
         "#]],
     );
 }
 
 #[test]
-fn use_of_dynamic_range_yields_no_errors() {
+fn use_of_dynamic_range_yields_errors() {
     check_profile(
         USE_DYNAMIC_RANGE,
         &expect![[r#"
-            []
+            [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 108,
+                        hi: 137,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 108,
+                        hi: 137,
+                    },
+                ),
+                UseOfDynamicRange(
+                    Span {
+                        lo: 108,
+                        hi: 137,
+                    },
+                ),
+            ]
         "#]],
     );
 }
 
 #[test]
-fn use_of_dynamic_double_yields_error() {
+fn use_of_dynamic_double_yields_errors() {
     check_profile(
         USE_DYNAMIC_DOUBLE,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 246,
+                        hi: 284,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 246,
+                        hi: 284,
+                    },
+                ),
                 UseOfDynamicDouble(
                     Span {
                         lo: 246,
@@ -100,6 +155,12 @@ fn use_of_dynamic_qubit_yields_errors() {
         USE_DYNAMIC_QUBIT,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 105,
+                        hi: 123,
+                    },
+                ),
                 UseOfDynamicQubit(
                     Span {
                         lo: 146,
@@ -117,6 +178,18 @@ fn use_of_dynamic_big_int_yields_errors() {
         USE_DYNAMIC_BIG_INT,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 247,
+                        hi: 285,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 247,
+                        hi: 285,
+                    },
+                ),
                 UseOfDynamicBigInt(
                     Span {
                         lo: 247,
@@ -134,6 +207,12 @@ fn use_of_dynamic_string_yields_errors() {
         USE_DYNAMIC_STRING,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 130,
+                        hi: 144,
+                    },
+                ),
                 UseOfDynamicString(
                     Span {
                         lo: 130,
@@ -146,11 +225,23 @@ fn use_of_dynamic_string_yields_errors() {
 }
 
 #[test]
-fn use_of_dynamically_sized_array_yields_error() {
+fn use_of_dynamically_sized_array_yields_errors() {
     check_profile(
         USE_DYNAMICALLY_SIZED_ARRAY,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 104,
+                        hi: 136,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 104,
+                        hi: 136,
+                    },
+                ),
                 UseOfDynamicallySizedArray(
                     Span {
                         lo: 104,
@@ -168,6 +259,18 @@ fn use_of_dynamic_udt_yields_errors() {
         USE_DYNAMIC_UDT,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 283,
+                        hi: 335,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 283,
+                        hi: 335,
+                    },
+                ),
                 UseOfDynamicDouble(
                     Span {
                         lo: 283,
@@ -187,14 +290,17 @@ fn use_of_dynamic_udt_yields_errors() {
 
 #[test]
 fn use_of_dynamic_function_yields_errors() {
+    // In the case of if expressions, if either the condition or the blocks yield errors, the errors yielded by the
+    // whole if expression are not surfaced to avoid too much error churn.
+    // For this reason, the "use of dynamic function" error is not yielded in this case.
     check_profile(
         USE_DYNAMIC_FUNCTION,
         &expect![[r#"
             [
-                UseOfDynamicArrowFunction(
+                UseOfDynamicBool(
                     Span {
                         lo: 142,
-                        hi: 166,
+                        hi: 154,
                     },
                 ),
             ]
@@ -204,14 +310,17 @@ fn use_of_dynamic_function_yields_errors() {
 
 #[test]
 fn use_of_dynamic_operation_yields_errors() {
+    // In the case of if expressions, if either the condition or the blocks yield errors, the errors yielded by the
+    // whole if expression are not surfaced to avoid too much error churn.
+    // For this reason, the "use of dynamic operation" error is not yielded in this case.
     check_profile(
         USE_DYNAMIC_OPERATION,
         &expect![[r#"
             [
-                UseOfDynamicArrowOperation(
+                UseOfDynamicBool(
                     Span {
                         lo: 142,
-                        hi: 162,
+                        hi: 154,
                     },
                 ),
             ]
@@ -230,11 +339,23 @@ fn call_cyclic_function_with_classical_argument_yields_no_errors() {
 }
 
 #[test]
-fn call_cyclic_function_with_dynamic_argument_yields_error() {
+fn call_cyclic_function_with_dynamic_argument_yields_errors() {
     check_profile(
         CALL_TO_CICLYC_FUNCTION_WITH_DYNAMIC_ARGUMENT,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 211,
+                        hi: 243,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 211,
+                        hi: 243,
+                    },
+                ),
                 CallToCyclicFunctionWithDynamicArg(
                     Span {
                         lo: 211,
@@ -256,6 +377,12 @@ fn call_cyclic_operation_with_classical_argument_yields_errors() {
                     Span {
                         lo: 15,
                         hi: 23,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 187,
+                        hi: 199,
                     },
                 ),
                 CallToCyclicOperation(
@@ -281,6 +408,18 @@ fn call_cyclic_operation_with_dynamic_argument_yields_errors() {
                         hi: 23,
                     },
                 ),
+                UseOfDynamicBool(
+                    Span {
+                        lo: 212,
+                        hi: 244,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 212,
+                        hi: 244,
+                    },
+                ),
                 CallToCyclicOperation(
                     Span {
                         lo: 212,
@@ -298,10 +437,16 @@ fn call_to_dynamic_function_yields_errors() {
         CALL_DYNAMIC_FUNCTION,
         &expect![[r#"
             [
-                UseOfDynamicArrowFunction(
+                UseOfDynamicBool(
                     Span {
                         lo: 142,
-                        hi: 166,
+                        hi: 154,
+                    },
+                ),
+                UseOfDynamicBool(
+                    Span {
+                        lo: 180,
+                        hi: 188,
                     },
                 ),
                 UseOfDynamicDouble(
@@ -333,10 +478,16 @@ fn call_to_dynamic_operation_yields_errors() {
         CALL_DYNAMIC_OPERATION,
         &expect![[r#"
             [
-                UseOfDynamicArrowOperation(
+                UseOfDynamicBool(
                     Span {
                         lo: 142,
-                        hi: 162,
+                        hi: 154,
+                    },
+                ),
+                UseOfDynamicBool(
+                    Span {
+                        lo: 176,
+                        hi: 181,
                     },
                 ),
                 UseOfDynamicArrowOperation(
@@ -380,11 +531,24 @@ fn call_to_unresolved_yields_errors() {
 }
 
 #[test]
-fn measurement_within_dynamic_scope_yields_no_errors() {
+fn measurement_within_dynamic_scope_yields_errors() {
     check_profile(
         MEASUREMENT_WITHIN_DYNAMIC_SCOPE,
         &expect![[r#"
-            []
+            [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 99,
+                        hi: 110,
+                    },
+                ),
+                MeasurementWithinDynamicScope(
+                    Span {
+                        lo: 137,
+                        hi: 141,
+                    },
+                ),
+            ]
         "#]],
     );
 }
@@ -395,6 +559,30 @@ fn use_of_dynamic_index_yields_errors() {
         USE_DYNAMIC_INDEX,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 246,
+                        hi: 271,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 246,
+                        hi: 271,
+                    },
+                ),
+                UseOfDynamicBool(
+                    Span {
+                        lo: 319,
+                        hi: 323,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 319,
+                        hi: 323,
+                    },
+                ),
                 UseOfDynamicIndex(
                     Span {
                         lo: 319,
@@ -407,11 +595,24 @@ fn use_of_dynamic_index_yields_errors() {
 }
 
 #[test]
-fn return_within_dynamic_scope_yields_no_errors() {
+fn return_within_dynamic_scope_yields_errors() {
     check_profile(
         RETURN_WITHIN_DYNAMIC_SCOPE,
         &expect![[r#"
-            []
+            [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 98,
+                        hi: 109,
+                    },
+                ),
+                ReturnWithinDynamicScope(
+                    Span {
+                        lo: 128,
+                        hi: 136,
+                    },
+                ),
+            ]
         "#]],
     );
 }
@@ -422,10 +623,52 @@ fn loop_with_dynamic_condition_yields_errors() {
         LOOP_WITH_DYNAMIC_CONDITION,
         &expect![[r#"
             [
+                UseOfDynamicBool(
+                    Span {
+                        lo: 106,
+                        hi: 118,
+                    },
+                ),
+                UseOfDynamicBool(
+                    Span {
+                        lo: 141,
+                        hi: 159,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 141,
+                        hi: 159,
+                    },
+                ),
+                UseOfDynamicRange(
+                    Span {
+                        lo: 141,
+                        hi: 159,
+                    },
+                ),
                 LoopWithDynamicCondition(
                     Span {
                         lo: 141,
                         hi: 159,
+                    },
+                ),
+                UseOfDynamicBool(
+                    Span {
+                        lo: 150,
+                        hi: 156,
+                    },
+                ),
+                UseOfDynamicInt(
+                    Span {
+                        lo: 150,
+                        hi: 156,
+                    },
+                ),
+                UseOfDynamicRange(
+                    Span {
+                        lo: 150,
+                        hi: 156,
                     },
                 ),
             ]
