+++ compiler/qsc_rca/tests/callables.rs
@@ -180,19 +180,19 @@ fn check_rca_for_unrestricted_h() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -278,25 +278,25 @@ fn check_rca_for_unrestricted_r1() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -394,25 +394,25 @@ fn check_rca_for_unrestricted_rx() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -516,31 +516,31 @@ fn check_rca_for_unrestricted_rxx() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                         [2]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                         [2]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -650,25 +650,25 @@ fn check_rca_for_unrestricted_ry() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -772,31 +772,31 @@ fn check_rca_for_unrestricted_ryy() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                         [2]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                         [2]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -906,25 +906,25 @@ fn check_rca_for_unrestricted_rz() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -1028,31 +1028,31 @@ fn check_rca_for_unrestricted_rzz() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                         [2]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicDouble)
                             value_kind: Element(Static)
                         [1]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                         [2]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -1156,19 +1156,19 @@ fn check_rca_for_unrestricted_s() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -1248,19 +1248,19 @@ fn check_rca_for_unrestricted_t() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -1340,19 +1340,19 @@ fn check_rca_for_unrestricted_x() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -1432,19 +1432,19 @@ fn check_rca_for_unrestricted_y() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
@@ -1524,19 +1524,19 @@ fn check_rca_for_unrestricted_z() {
                             value_kind: Element(Static)
                 ctl: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)
                 ctl-adj: ApplicationsGeneratorSet:
                     inherent: Quantum: QuantumProperties:
-                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | ForwardBranchingOnDynamicValue)
+                        runtime_features: RuntimeFeatureFlags(UseOfDynamicBool)
                         value_kind: Element(Static)
                     dynamic_param_applications:
                         [0]: [Parameter Type Element] Quantum: QuantumProperties:
-                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit | ForwardBranchingOnDynamicValue)
+                            runtime_features: RuntimeFeatureFlags(UseOfDynamicBool | UseOfDynamicQubit)
                             value_kind: Element(Static)"#
         ],
     );
